lab/PoC: OCP CI Sippy querying e2e flakes
This PoC will query the Sippy API looking for flakes
given an e2e test from stdin.

This is useful when looking if specific tests are flaking
on CI, and how often (percentage).

The query to Sippy API is serial, so may not impact the backend.

Example running
```bash
$ cat sample-test.txt  | ./ocp-ci-query-flakes
Flakes	Perc	 TestName
98	10.262%	[sig-arch][bz-DNS][Late] Alerts alert/KubePodNotReady should not be at or above pending in ns/openshift-dns [Suite:openshift/conformance/parallel]
```